refactor: remove duplicate components / fix typos
