+++ packages/front-end/.eslintrc.json
@@ -1,7 +1,6 @@
 {
   "extends": [
     "next/core-web-vitals",
-    "plugin:@tanstack/eslint-plugin-query/recommended",
-    "plugin:storybook/recommended"
+    "plugin:@tanstack/eslint-plugin-query/recommended"
   ]
 }
